requirements.txt
tensorflow-cpu==2.12.0
pandas
numpy
scikit-learn
tensorboard
matplotlib
streamlit